requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sse-starlette>=2.0.0
gspread>=6.0.0
google-auth>=2.25.0
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
pdf2image>=1.17.0
Pillow>=10.0.0
streamlit>=1.30.0
